♻️ 모달 컴포넌트 통합 및 리팩토링
